AllGoals and SingleGoal
Changed layouts of All Goals page and Single Goal page. All Goals now
has a listView with entries containing a color based on the reason, the
goal's name, the goal's reason, and 2 progress bars for total progress
and the current month's progress. SingleGoal page has the same
information, and also contains a calendar showing the current goal's
events for that month.